# src/tsoa-generated/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"NPMPackage": {
				"description": "The result of a package request against `https://registry.npmjs.org`. This is\na subset of the returned data, not a full representation, that contains\neverything you will need to carry out the exercise.",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dist-tags": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					},
					"versions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"version": {
						"type": "string"
					},
					"dependencies": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"name",
					"description",
					"dist-tags",
					"versions",
					"version"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NPMPackageComparison": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dist-tags": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					},
					"versions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"version": {
						"type": "string"
					},
					"dependencies": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					},
					"dependencyDifferences": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"name",
					"description",
					"dist-tags",
					"versions",
					"version",
					"dependencyDifferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NPMPackageParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"version"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "dependency-problem",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/dependency/{name}/{version}": {
			"get": {
				"operationId": "GetDependency",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/NPMPackage"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "version",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "GetDependencyMismatches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/NPMPackage"
										},
										{
											"$ref": "#/components/schemas/NPMPackageComparison"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "version",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/NPMPackageParams"
								},
								"type": "array"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}